fdp-class

fdp-class

fdp-class

Details

Container for class fdp

Public fields

yaml

a list containing the contents of the working config.yaml

fdp_config_dir

a string specifying the directory passed from fair run

model_config

a string specifying the URL of an entry in the object table associated with the storage_location of the working config.yaml

submission_script

a string specifying the URL of an entry in the object table associated with the storage_location of the submission script

code_repo

a string specifying the URL of an entry in the object table associated with the GitHub repository

code_run

a string specifying the URL of an entry in the code_run table

inputs

a data.frame containing metadata associated with code_run inputs

outputs

a data.frame containing metadata associated with code_run outputs

issues

a data.frame containing metadata associated with code_run issues

Methods

Public methods

fdp$new()

fdp$print()

fdp$input()

fdp$output()

fdp$output_index()

fdp$raise_issue()

fdp$finalise_output_hash()

fdp$finalise_output_url()

fdp$clone()

Method new()

Create a new fdp object

Usage

fdp$new(
  yaml,
  fdp_config_dir,
  model_config,
  submission_script,
  code_repo,
  code_run
)

Arguments

yaml

a list containing the contents of the working config.yaml

fdp_config_dir

a string specifying the directory passed from fair run

model_config

a string specifying the URL of an entry in the object table associated with the storage_location of the working config.yaml

submission_script

a string specifying the URL of an entry in the object table associated with the storage_location of the submission script

code_repo

a string specifying the URL of an entry in the object table associated with the GitHub repository

code_run

a string specifying the URL of an entry in the code_run table

Returns

Returns a new fdp object

Method print()

Print method

Usage

fdp$print(...)

Method input()

Record code_run inputs in fdp object

Usage

fdp$input(
  data_product,
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  path,
  component_url
)

Arguments

data_product

a string specifying the name of the data product, used as a reference

use_data_product

a string specifying the name of the data product, used as input in the code_run

use_component

a string specifying the name of the data product component, used as input in the code_run

use_version

a string specifying the data product version, used as input in the code_run

use_namespace

a string specifying the namespace in which the data product resides, used as input in the code_run

path

a string specifying the location of the data product in the local data store

component_url

a string specifying the URL of an entry in the object_component table

Returns

Returns an updated fdp object

Method output()

Record code_run outputs in fdp object

Usage

fdp$output(
  data_product,
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  path,
  data_product_description,
  component_description,
  public
)

Arguments

data_product

a string specifying the name of the data product, used as a reference

use_data_product

a string specifying the name of the data product, used as output in the code_run

use_component

a string specifying the name of the data product component, used as output in the code_run

use_version

a string specifying the version of the data product, used as output in the code_run

use_namespace

a string specifying the namespace in which the data product resides, used as output in the code_run

path

a string specifying the location of the data product in the local data store

data_product_description

a string containing a description of the data product

component_description

a string containing a description of the data product component

public

Returns

Returns an updated fdp object

Method output_index()

Return index of data product recorded in fdp object so that an issue may be attached

Usage

fdp$output_index(data_product, component, version, namespace)

Arguments

data_product

a string specifying the name of the data product, used as output in the code_run

component

a string specifying the name of the data product component, used as output in the code_run

version

a string specifying the name of the data product version, used as output in the code_run

namespace

a string specifying the namespace in which the data product resides, used as input in the code_run

Returns

Returns an index used to identify the data product

Method raise_issue()

Record issue in fdp object

Usage

fdp$raise_issue(
  index,
  type,
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  issue,
  severity
)

Arguments

index

a numeric index, used to identify each input and output in the fdp object

type

a string specifying the type of issue (one of "data", "config", "script", "repo")

use_data_product

a string specifying the name of the data product, used as output in the code_run

use_component

a string specifying the name of the data product component, used as output in the code_run

use_version

a string specifying the name of the data product version, used as output in the code_run

use_namespace

a string specifying the namespace in which the data product resides, used as input in the code_run

issue

a string containing a free text description of the issue

severity

an integer specifying the severity of the issue

Returns

Returns an updated fdp object

Method finalise_output_hash()

Record file hash and update path name in fdp object

Usage

fdp$finalise_output_hash(
  use_data_product,
  use_data_product_runid,
  use_version,
  use_namespace,
  hash,
  new_path,
  data_product_url,
  delete_if_duplicate = FALSE
)

Arguments

use_data_product

a string specifying the name of the data product, used as output in the code_run

use_data_product_runid

a string specifying the name of the data product, the same as use_data_product excluding the RUN_ID variable

use_version

a string specifying the name of the data product version, used as output in the code_run

use_namespace

a string specifying the namespace in which the data product resides, used as input in the code_run

hash

a string specifying the hash of the file

new_path

a string specifying the updated location (filename is now the hash of the file) of the data product in the local data store

data_product_url

a string specifying the URL of an object associated with the data_product

delete_if_duplicate

(optional) default is FALSE

Returns

Returns an updated fdp object

Method finalise_output_url()

Record data_product and component URLs in fdp object

Usage

fdp$finalise_output_url(
  use_data_product,
  use_component,
  use_version,
  use_namespace,
  component_url
)

Arguments

use_data_product

a string specifying the name of the data product, used as output in the code_run

use_component

a string specifying the name of the data product component, used as output in the code_run

use_version

a string specifying the name of the data product version, used as output in the code_run

use_namespace

a string specifying the namespace in which the data product resides, used as input in the code_run

component_url

a string specifying the URL of an entry in the object_component table

Returns

Returns an updated fdp object

Method clone()

The objects of this class are cloneable with this method.

Usage

fdp$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.