Get current GitHub hash

Get the hash of the latest commit in the master branch of a particular repository. This function assumes git is installed and located in the System PATH.

get_github_hash(repo)

Arguments

repo

a string specifying the github username/repository

See also