Post entry to namespace table

Upload information to the namespace table in the data registry

new_namespace(
  name,
  full_name,
  website,
  endpoint = "http://localhost:8000/api/"
)

Arguments

name
a string specifying the name of the namespace

full_name
(optional) a string specifying the full name of the namespace

website
(optional) a string specifying the website URL associated with the namespace

endpoint
a string specifying the registry endpoint

See also